+++ python/ml4ir/applications/ranking/tests/test_fixed_additive_positional_bias.py
@@ -5,6 +5,7 @@ from ml4ir.base.model.architectures.fixed_additive_positional_bias import FixedA
 
 warnings.filterwarnings("ignore")
 
+INPUT_DIR = "/ml4ir/applications/ranking/tests/data/configs"
 
 class TestFixedAdditivePositionalBias(unittest.TestCase):
 
@@ -25,6 +26,21 @@ class TestFixedAdditivePositionalBias(unittest.TestCase):
         self.apply_additive_positional_bias([1,2], 2, True)
         self.apply_additive_positional_bias([2], 2, False)
         self.apply_additive_positional_bias([2,4,6,8], 10, True)
+
+    def test_zeros_weight_initialization(self):
+        """Testing weight initializations to zeros"""
+        positional_bias = FixedAdditivePositionalBias(max_ranks=5, kernel_initializer='Zeros')
+        positional_bias(tf.constant([1.]))
+        weights = positional_bias.dense.get_weights()
+        assert all([w == 0. for w in weights[0]])
+
+    def test_non_zeros_weight_initialization(self):
+        """Testing weight initializations to non zeros"""
+        positional_bias = FixedAdditivePositionalBias(max_ranks=5, kernel_initializer='glorot_uniform')
+        positional_bias(tf.constant([1.]))
+        weights = positional_bias.dense.get_weights()
+        assert any([w != 0. for w in weights[0]])
+
 
 
 if __name__ == "__main__":