+++ python/ml4ir/base/pipeline.py
@@ -137,6 +137,10 @@ class RelevancePipeline(object):
         # Set random seeds
         self.set_seeds()
 
+        self.logger.info("Running pre-processing step.")
+        self.pre_processing_step()
+        self.logger.info("Pre-processing step done.")
+
         # Read/Parse feature_config and model_config YAML
         feature_config_dict = self.file_io.read_yaml(args.feature_config)
         model_config_dict = self.file_io.read_yaml(args.model_config)
@@ -282,8 +286,6 @@ class RelevancePipeline(object):
             job_info = ""
             train_metrics = dict()
             test_metrics = dict()
-
-            self.pre_processing_step()
 
             # Build dataset
             relevance_dataset = self.get_relevance_dataset()
@@ -444,6 +446,14 @@ class RelevancePipeline(object):
         """
         return self
 
+    def post_training_step(self):
+        """
+        Performs arbitrary post-training steps such as copying or transforming data that the rest of the code can not
+        accommodate. It serves as a placeholder without an explicit implementation (returns self) in the base pipeline.
+        We expect that users can extend it in their custom pipelines.
+        """
+        return self
+
     def finish(self, job_status, job_info):
         """
         Wrap up the model training pipeline.
@@ -478,6 +488,10 @@ class RelevancePipeline(object):
                 self.models_dir_local, self.models_dir, overwrite=True)
             self.file_io.copy_to_hdfs(
                 self.logs_dir_local, self.logs_dir, overwrite=True)
+
+        self.logger.info("Running post-training step.")
+        self.post_training_step()
+        self.logger.info("Post-training step done.")
 
         e = int(time.time() - self.start_time)
         self.logger.info(